proyect/requirements.txt
Django
guinicorn
djangorestframework
django-cors-headers
antlr-tools
antlr4-python3-runtime==4.13.2